+++ src/main/java/roomescape/controller/api/MemberApiController.java
@@ -2,7 +2,6 @@ package roomescape.controller.api;
 
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,7 +11,6 @@ import roomescape.service.dto.response.MemberIdAndNameResponse;
 import roomescape.service.member.MemberService;
 
 import java.net.URI;
-import java.util.List;
 
 @RestController
 public class MemberApiController {
@@ -21,16 +19,6 @@ public class MemberApiController {
 
     public MemberApiController(MemberService memberService) {
         this.memberService = memberService;
-    }
-
-    @GetMapping("/admin/members")
-    public ResponseEntity<List<MemberIdAndNameResponse>> getMembers() {
-        List<Member> members = memberService.findMembers();
-        return ResponseEntity.ok(
-                members.stream()
-                        .map(member -> new MemberIdAndNameResponse(member.getId(), member.getName()))
-                        .toList()
-        );
     }
 
     @PostMapping("/members")
